+++ lib/mcptask_rails_runner/binary.rb
@@ -10,7 +10,7 @@ module McptaskRailsRunner
     SUPPORTED_PLATFORMS = %w[
       arm64-darwin
       x86_64-darwin
-      arm64-linux
+      aarch64-linux
       x86_64-linux
       x64-mingw-ucrt
     ].freeze